// apps/game-client/src/lib/api/generated/main/model/update-sound-settings-dto.ts
/**
 * Generated by orval v8.8.1 🍺
 * Do not edit manually.
 * Lootlog API
 * The Lootlog API documentation
 * OpenAPI spec version: 1.0
 */
import type { UpdateSoundSettingsDtoDetectorConfig } from './update-sound-settings-dto-detector-config';
import type { UpdateSoundSettingsDtoNotificationsConfig } from './update-sound-settings-dto-notifications-config';
import type { UpdateSoundSettingsDtoTimersConfig } from './update-sound-settings-dto-timers-config';

export interface UpdateSoundSettingsDto {
  /**
     * @minimum 0
     * @maximum 1
     */
  masterVolume?: number;
  /**
     * @minimum 0
     * @maximum 1
     */
  notificationsVolume?: number;
  /**
     * @minimum 0
     * @maximum 1
     */
  detectorVolume?: number;
  /**
     * @minimum 0
     * @maximum 1
     */
  timersVolume?: number;
  notificationsConfig?: UpdateSoundSettingsDtoNotificationsConfig;
  detectorConfig?: UpdateSoundSettingsDtoDetectorConfig;
  timersConfig?: UpdateSoundSettingsDtoTimersConfig;
}